Moved extensions to separate namespace
